added judge_supervise to admin for judges
